moved DRequest and DResult into protocol
