+++ distributed/src/main/scala/asobu/distributed/gateway/package.scala
@@ -1,5 +1,6 @@
 package asobu.distributed
 
+import asobu.distributed.protocol.DRequest
 import asobu.dsl.Extractor
 
 package object gateway {
