+++ distributed/src/main/scala/asobu/distributed/service/extractors/package.scala
@@ -1,6 +1,6 @@
 package asobu.distributed.service
 
-import asobu.distributed.DRequest
+import asobu.distributed.protocol.DRequest
 import asobu.dsl.Extractor
 
 package object extractors {
